move common func back to auth from pro/auth
